STYLE: Remove outdated cmake variable
Old executable location variable.

The target property RUNTIME_OUTPUT_DIRECTORY supercedes this variable
for a target if it is set. Executable targets are otherwise placed in
this directory.

Old library location variable.

The target properties ARCHIVE_OUTPUT_DIRECTORY,
LIBRARY_OUTPUT_DIRECTORY, and RUNTIME_OUTPUT_DIRECTORY supercede
this variable for a target if they are set. Library targets are
otherwise placed in this directory.